# Conflicts:
#	eth-withdrawal/src/main/kotlin/withdrawalservice/WithdrawalServiceImpl.kt
#	eth/src/main/kotlin/notary/endpoint/eth/EthRefundStrategyImpl.kt
#	notary-btc-integration-test/src/main/kotlin/integration/helper/BtcIntegrationHelperUtil.kt
#	notary-commons/src/main/kotlin/registration/IrohaAccountCreator.kt
#	notary-commons/src/main/kotlin/registration/RegistrationServiceEndpoint.kt
#	notary-commons/src/main/kotlin/sidechain/iroha/util/ModelUtil.kt
#	notary-commons/src/main/kotlin/sidechain/iroha/util/RequestHelper.kt
#	notary-commons/src/test/java/registration/IrohaAccountCreatorTest.java
#	notary-eth-integration-test/src/main/kotlin/integration/helper/EthIntegrationHelperUtil.kt
#	notary-sora-integration-test/src/integration-test/kotlin/integration/sora/SoraIntegrationTest.kt

+++ eth-withdrawal/src/main/kotlin/withdrawalservice/WithdrawalServiceImpl.kt
@@ -153,7 +153,7 @@ class WithdrawalServiceImpl(
                 }
 
                 val (coinAddress, precision) = tokenInfo
-                val decimalAmount = BigDecimal(amount).scaleByPowerOfTen(precision.toInt()).toPlainString()
+                val decimalAmount = BigDecimal(amount).scaleByPowerOfTen(precision).toPlainString()
                 RollbackApproval(coinAddress, decimalAmount, address, hash, rr, ss, vv, relayAddress)
             }
     }
